KDKPAS-3408 Remove status command, and add polling to get-report command
It was decided that we don't want to have a single function that does
everything and deletes the transfer with a single command, instead we
have separate commands for the different functionalities. Though polling
was added as a part of the report downloading command, so that the user
doesn't have to manually check when the transfer is complete.